webserver/nginx/requirements.txt
install==1.3.5
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
